ai-business-management/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
sqlalchemy>=2.0.36
joblib>=1.4.0
scikit-learn>=1.6.0
numpy>=1.26.0
pandas>=2.2.0
python-dotenv>=1.0.0
pymysql>=1.1.0
cryptography>=43.0.0
gspread>=6.0.0
google-auth>=2.27.0
